Created and added a new Thread policy. Also updated the thread view to use that policy and only allows thread creators to delete if thread belongs to them.

+++ app/Http/Controllers/ThreadController.php
@@ -124,11 +124,18 @@ class ThreadController extends Controller
     public function destroy($channel, Thread $thread)
     {
 //        $thread->replies()->delete();
+
+//        if ($thread->user_id != auth()->id()) {
+//            abort(403, 'Permission Denied');
+//        }
+
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if (request()->wantsJson()) {
             //        return response([], 204);
-            return response([], 302);
+            return response([], 204);
         }
         return redirect('/threads');
